+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 11.1.0
+
+- Added instruction manual support
+- Requires Assorted Lib 4.1.0.
+
 ## 11.0.0
 
 Updated to Minecraft 26.2, for NeoForge and Fabric.